Set timeout to a higher value
The default read timeout for httpx is 5 seconds, it is not enough
for some long running chat completion requests.

Therefore, I set the default timeout value to something higher.

Note that, previously, when using requests library, we were using
the default timeout, which was infinite. I set this new timeout to
10 minutes, which should be enough almost all the time.